pyproject.toml
[project]
name = "parsemedicalexams"
dynamic = ["version"]
description = "Extract and summarize medical exam reports (X-rays, MRIs, ultrasounds, etc.) with AI precision"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    {name = "Tiago Silva", email = "tiago@tsilva.com"}
]
keywords = ["medical", "exam-parser", "ai", "document-processing", "radiology"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "openai",
    "pillow",
    "pdf2image",
    "python-dotenv",
    "pydantic",
    "pyyaml",
    "tqdm",
    "typing-extensions",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-cov",
    "ruff",
    "mypy",
    "types-PyYAML",
    "types-tqdm",
]

[project.scripts]
medicalexamsparser = "parsemedicalexams.cli:main"
parsemedicalexams = "parsemedicalexams.cli:main"

[project.urls]
Homepage = "https://github.com/tsilva/parsemedicalexams"
Repository = "https://github.com/tsilva/parsemedicalexams"
Issues = "https://github.com/tsilva/parsemedicalexams/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.version]
path = "parsemedicalexams/__init__.py"

[tool.ruff]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]
